refactor: move cluster data classes out of network package
