+++ driver/src/main/java/eu/cloudnetservice/driver/channel/ChannelMessageSender.java
@@ -18,8 +18,8 @@ package eu.cloudnetservice.driver.channel;
 
 import eu.cloudnetservice.driver.ComponentInfo;
 import eu.cloudnetservice.driver.DriverEnvironment;
+import eu.cloudnetservice.driver.cluster.NetworkClusterNode;
 import eu.cloudnetservice.driver.inject.InjectionLayer;
-import eu.cloudnetservice.driver.network.cluster.NetworkClusterNode;
 import eu.cloudnetservice.driver.service.ServiceInfoSnapshot;
 import lombok.NonNull;
