+++ modules/rest/src/main/java/eu/cloudnetservice/modules/rest/v2/V2HttpHandlerCluster.java
@@ -17,7 +17,7 @@
 package eu.cloudnetservice.modules.rest.v2;
 
 import eu.cloudnetservice.common.document.gson.JsonDocument;
-import eu.cloudnetservice.driver.network.cluster.NetworkClusterNode;
+import eu.cloudnetservice.driver.cluster.NetworkClusterNode;
 import eu.cloudnetservice.driver.network.http.HttpContext;
 import eu.cloudnetservice.driver.network.http.HttpResponseCode;
 import eu.cloudnetservice.driver.network.http.annotation.HttpRequestHandler;
