+++ node/src/main/java/eu/cloudnetservice/node/config/Configuration.java
@@ -17,9 +17,9 @@
 package eu.cloudnetservice.node.config;
 
 import eu.cloudnetservice.common.document.gson.JsonDocument;
+import eu.cloudnetservice.driver.cluster.NetworkCluster;
+import eu.cloudnetservice.driver.cluster.NetworkClusterNode;
 import eu.cloudnetservice.driver.network.HostAndPort;
-import eu.cloudnetservice.driver.network.cluster.NetworkCluster;
-import eu.cloudnetservice.driver.network.cluster.NetworkClusterNode;
 import eu.cloudnetservice.driver.network.ssl.SSLConfiguration;
 import jakarta.inject.Singleton;
 import java.util.Collection;
