+++ node/src/main/java/eu/cloudnetservice/node/network/listener/PacketClientServiceSyncAckListener.java
@@ -16,8 +16,8 @@
 
 package eu.cloudnetservice.node.network.listener;
 
+import eu.cloudnetservice.driver.cluster.NodeInfoSnapshot;
 import eu.cloudnetservice.driver.network.NetworkChannel;
-import eu.cloudnetservice.driver.network.cluster.NodeInfoSnapshot;
 import eu.cloudnetservice.driver.network.def.NetworkConstants;
 import eu.cloudnetservice.driver.network.protocol.Packet;
 import eu.cloudnetservice.driver.network.protocol.PacketListener;
