+++ node/src/main/java/eu/cloudnetservice/node/network/packet/PacketServerServiceSyncAckPacket.java
@@ -16,8 +16,8 @@
 
 package eu.cloudnetservice.node.network.packet;
 
+import eu.cloudnetservice.driver.cluster.NodeInfoSnapshot;
 import eu.cloudnetservice.driver.network.buffer.DataBuf;
-import eu.cloudnetservice.driver.network.cluster.NodeInfoSnapshot;
 import eu.cloudnetservice.driver.network.def.NetworkConstants;
 import eu.cloudnetservice.driver.network.protocol.BasePacket;
 import lombok.NonNull;
